proxy/pyproject.toml
[project]
name = "proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.3",
    "geoip2>=5.1.0",
    "httpx>=0.28.1",
    "logging>=0.4.9.6",
    "opa-python-client>=2.0.3",
    "python-dotenv>=1.1.1",
    "python-keycloak>=5.8.1",
    "uvicorn>=0.37.0",
]
